pyproject.toml
[build-system]
requires = ['flit_core>=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'aiida-amcsd'
dynamic = ['version']
description = 'AiiDA package to import and analyze data from American Mineralogist Crystal Structure Database.'
authors = [
    {name = 'Sebastiaan P. Huber', email = 'mail@sphuber.net'}
]
readme = 'README.md'
license = {file = 'LICENSE.txt'}
classifiers = [
    'Development Status :: 3 - Alpha',
    'Framework :: AiiDA',
    'License :: OSI Approved :: MIT License',
    'Operating System :: POSIX :: Linux',
    'Operating System :: MacOS :: MacOS X',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Topic :: Scientific/Engineering'
]
keywords = ['aiida', 'AMCSD']
requires-python = '>=3.8'
dependencies = [
    'aiida-core~=2.0',
    'beautifulsoup4~=4.0',
    'click~=8.0',
    'pycifrw~=4.4',
    'requests~=2.0',
    'requests-cache~=0.9',
]

[project.urls]
Source = 'https://github.com/sphuber/aiida-amcsd'

[project.scripts]
aiida-amcsd = 'aiida_amcsd.cli.cli:cli'

[project.entry-points.'aiida.calculations']
'amcsd.cif_file_to_cif_data' = 'aiida_amcsd.calculations.functions.cif_file_to_cif_data:cif_file_to_cif_data'

[project.optional-dependencies]
pre-commit = [
    'pre-commit~=2.17',
    'pylint~=2.12.2',
    'pylint-aiida~=0.1.1',
]
tests = [
    'pgtest~=1.3,>=1.3.1',
    'pytest~=6.2',
]

[tool.flit.module]
name = 'aiida_amcsd'

[tool.flit.sdist]
exclude = [
    'tests/',
]

[tool.flynt]
line-length = 120
fail-on-change = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.pydocstyle]
ignore = [
    'D104',
    'D203',
    'D213'
]

[tool.pylint.master]
load-plugins = ['pylint_aiida']

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    'bad-continuation',
    'duplicate-code',
    'import-outside-toplevel',
    'too-many-arguments',
    'too-many-ancestors',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 120
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true
